P4-2185 - importing reporting files
Add sample reporting files both for localstack and for test resources.

Move ReportingImporter.kt to ReportingParser.kt, and create a new ReportingImporter.kt that gets reporting content using a Reporting Provider.
