+++ localstack/2020-09-01-people.jsonl
@@ -1,0 +1,1 @@
+{"id":"PE1","created_at":"2020-09-07T16:25:15+01:00","updated_at":"2020-09-07T16:25:24+01:00","criminal_records_office":null,"nomis_prison_number":null,"police_national_computer":"PNC1","prison_number":"PRISON1","latest_nomis_booking_id":null,"gender":"male","age":46}
