+++ src/test/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/TestConfig.kt
@@ -8,6 +8,7 @@ import uk.gov.justice.digital.hmpps.pecs.jpc.config.GeoamyPricesProvider
 import uk.gov.justice.digital.hmpps.pecs.jpc.config.ReportingProvider
 import uk.gov.justice.digital.hmpps.pecs.jpc.config.Schedule34LocationsProvider
 import uk.gov.justice.digital.hmpps.pecs.jpc.config.SercoPricesProvider
+import uk.gov.justice.digital.hmpps.pecs.jpc.reporting.ReportingImporter
 import java.time.Clock
 
 @TestConfiguration
@@ -36,6 +37,9 @@ class TestConfig {
 
     @Bean
     fun reportingResourceProvider(): ReportingProvider {
-        return ReportingProvider { resourceLoader.getResource(it).file.readText() }
+        return ReportingProvider { resourceLoader.getResource("classpath:/reporting/$it").file.readText() }
     }
+
+    @Bean
+    fun reportImporter() = ReportingImporter()
 }
