+++ src/test/resources/reporting/2020-09-01-profiles.jsonl
@@ -1,0 +1,1 @@
+{"id":"PR1","person_id":"PE1"}
