+++ src/test/resources/reporting/2020-09-02-profiles.jsonl
@@ -1,0 +1,2 @@
+{"id":"PR2","person_id":"PE1"}
+{"id":"PR3","person_id":"PE2"}
