--- src/test/resources/reporting/2020-09-08-moves.jsonl
@@ -1,4 +1,0 @@
-{"id":"02b4c0f5-4d85-4fb6-be6c-53d74b85bf2e","date":"2021-02-28","status":"requested","created_at":"2020-09-07T15:30:48+01:00","updated_at":"2020-09-07T15:30:59+01:00","reference":"UKW4591N","move_type":"prison_transfer","additional_information":null,"time_due":null,"cancellation_reason":null,"cancellation_reason_comment":null,"profile_id":null,"reason_comment":null,"move_agreed":null,"move_agreed_by":null,"date_from":null,"date_to":null,"allocation_id":"e05ee488-33a2-489f-9e57-5cab1871e2a0","rejection_reason":null,"from_location_type":"prison","from_location":"WYI","to_location_type":"prison","to_location":"GNI","supplier":"geoamey"}
-{"id":"02d0f1cc-42d2-47e3-a09a-a50fcd02aec7","date":"2020-09-07","status":"requested","created_at":"2020-09-07T15:44:08+01:00","updated_at":"2020-09-07T15:44:08+01:00","reference":"VNJ9618P","move_type":"court_appearance","additional_information":null,"time_due":null,"cancellation_reason":null,"cancellation_reason_comment":null,"profile_id":"d4168c75-21da-4694-866e-8fae71f327e6","reason_comment":null,"move_agreed":null,"move_agreed_by":null,"date_from":null,"date_to":null,"allocation_id":null,"rejection_reason":null,"from_location_type":"prison","from_location":"WYI","to_location_type":"court","to_location":"BATHYC","supplier":"geoamey"}
-{"id":"03d24b53-1cbb-48a3-93e6-f447e0ec1ca1","date":"2020-09-12","status":"requested","created_at":"2020-09-07T12:44:19+01:00","updated_at":"2020-09-07T15:00:38+01:00","reference":"HTP4975Y","move_type":"prison_transfer","additional_information":null,"time_due":null,"cancellation_reason":null,"cancellation_reason_comment":null,"profile_id":null,"reason_comment":null,"move_agreed":null,"move_agreed_by":null,"date_from":null,"date_to":null,"allocation_id":"54f480c2-60ad-48fd-9213-1a3ed6380fdf","rejection_reason":null,"from_location_type":"prison","from_location":"TSI","to_location_type":"prison","to_location":"EXI","supplier":"serco"}
-{"id":"03fca39a-354f-4240-aae8-3d50e4ec6418","date":"2020-09-07","status":"cancelled","created_at":"2020-09-07T12:49:24+01:00","updated_at":"2020-09-07T12:50:59+01:00","reference":"UHC1374V","move_type":"court_appearance","additional_information":null,"time_due":null,"cancellation_reason":"made_in_error","cancellation_reason_comment":"","profile_id":"89baf91f-f51f-4a1f-b3ab-848d45e34891","reason_comment":null,"move_agreed":null,"move_agreed_by":null,"date_from":null,"date_to":null,"allocation_id":null,"rejection_reason":null,"from_location_type":"police","from_location":"CLP1","to_location_type":"court","to_location":"EASBYC","supplier":"serco"}
